Fixes clipboard sync (#1250)
* Now that clipboard.readText() requires transient user activation
(see https://developer.mozilla.org/en-US/docs/Web/API/Clipboard/readText#security),
our system of syncing the local clipboard to remote onfocus and onmouseenter
no longer works, since those events don't necessarily imply a transient event.

With this change, we instead sendLocalClipboardToRemote onMouseDown and onKeyDown,
which should guarantee that the sync is always within a transient event.
Because the user must preform at least one of these actions in order to
initiate a paste, the clipboard should appear to always be in sync from
their perspective.

* eliminates the entire useClipboard hook and its surrounding state, which
was a big rube goldberg machine ensuring that
1. If a user had clipboard sharing enabled by RBAC
2. they were on chrome and
3. they had clipboard permissions allowed

If 1 was true, but either of 2 or 3 wasn’t, we blocked the UI and forced
them to correct it. That was also the mechanism that caused the clipboard
permissions dialog to pop up at the beginning of the session if clipboard
access wasn’t "granted".

Now the UX is much simpler: if a user has clipboard sharing allowed by RBAC,
and they’re on Chrome, clipboard sharing is enabled. If either of those is false,
clipboard sharing is disabled. If a user’s clipboard permissions are in
"prompt" state (aka neither "granted" nor "denied"), the first time they
try to copy/cut something on the remote, or have something in their clipboard
locally and click a mouse button or press a key, the browser prompt asking
for clipboard permissions will show up.